refactor: decompose CLI cmd/run.rs into focused sub-modules [OPE-287]
Split the 817-line monolithic run.rs into a module directory with clear
ownership boundaries:

- gateways.rs (93 LOC): GatewaySpec registry, collect_gateways()
- runtime.rs (163 LOC): All spawn_* functions, RetentionPolicy, cleanup
- tests.rs (435 LOC): Full test suite (14 tests)
- mod.rs (159 LOC): execute() orchestrator + module wiring

Zero behavior change. All 14 run module tests pass, clippy clean,
cargo fmt verified.
